+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Notable changes to OgbujiPT. Format based on [Keep a Changelog](https://keepacha
 ### Added
 
 - Implemented `ogbujipt.llm_wrapper.llama_cpp_http_chat` & `ogbujipt.llm_wrapper.llama_cpp_http`; llama.cpp low-level HTTP API support
+- Implemented flexible `ogbujipt.llm_wrapper.llama_response` class
+
+### Changed
+
+- Deprecated `first_choice_text` & `first_choice_message` methods in favor of `first_choice_text` attributes on response objects
 
 ## [0.7.1] - 20240229